creacion de los productos

+++ src/main/kotlin/entities/Products.kt
@@ -1,4 +1,4 @@
 package entities
 
 import java.util.*
-data class Products(var id: UUID = UUID.randomUUID(), var name: String, var description: String, var price: Int, var taxes: Int, var stock: Int)
+data class Products(var id: UUID = UUID.randomUUID(), var name: String, var description: String, var price: Float, var taxes: Int, var stock: Int)
